+++ banyand/metadata/schema/group.go
@@ -19,20 +19,16 @@ package schema
 
 import (
 	"context"
-	"strings"
+	"path"
 
 	"github.com/pkg/errors"
 	clientv3 "go.etcd.io/etcd/client/v3"
-	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 )
 
-var (
-	groupsKeyPrefix  = "/groups/"
-	groupMetadataKey = "/__meta_group__"
-)
+var groupsKeyPrefix = "/groups/"
 
 func (e *etcdSchemaRegistry) GetGroup(ctx context.Context, group string) (*commonv1.Group, error) {
 	var entity commonv1.Group
@@ -44,24 +40,15 @@ func (e *etcdSchemaRegistry) GetGroup(ctx context.Context, group string) (*commo
 }
 
 func (e *etcdSchemaRegistry) ListGroup(ctx context.Context) ([]*commonv1.Group, error) {
-	messages, err := e.client.Get(ctx, groupsKeyPrefix, clientv3.WithFromKey(), clientv3.WithRange(incrementLastByte(groupsKeyPrefix)))
+	messages, err := e.listWithPrefix(ctx, groupsKeyPrefix, KindGroup)
 	if err != nil {
 		return nil, err
 	}
-
-	var groups []*commonv1.Group
-	for _, kv := range messages.Kvs {
-		// kv.Key = "/groups/" + {group} + "/__meta_info__"
-		if strings.HasSuffix(string(kv.Key), groupMetadataKey) {
-			message := &commonv1.Group{}
-			if innerErr := proto.Unmarshal(kv.Value, message); innerErr != nil {
-				return nil, innerErr
-			}
-			groups = append(groups, message)
-		}
+	entities := make([]*commonv1.Group, 0, len(messages))
+	for _, message := range messages {
+		entities = append(entities, message.(*commonv1.Group))
 	}
-
-	return groups, nil
+	return entities, nil
 }
 
 func (e *etcdSchemaRegistry) DeleteGroup(ctx context.Context, group string) (bool, error) {
@@ -72,9 +59,9 @@ func (e *etcdSchemaRegistry) DeleteGroup(ctx context.Context, group string) (boo
 	keysToDelete := allKeys()
 	deleteOPs := make([]clientv3.Op, 0, len(keysToDelete)+1)
 	for _, key := range keysToDelete {
-		deleteOPs = append(deleteOPs, clientv3.OpDelete(listPrefixesForEntity(group, key), clientv3.WithPrefix()))
+		deleteOPs = append(deleteOPs, clientv3.OpDelete(e.prependNamespace(listPrefixesForEntity(group, key)), clientv3.WithPrefix()))
 	}
-	deleteOPs = append(deleteOPs, clientv3.OpDelete(formatGroupKey(group), clientv3.WithPrefix()))
+	deleteOPs = append(deleteOPs, clientv3.OpDelete(e.prependNamespace(formatGroupKey(group)), clientv3.WithPrefix()))
 	txnResponse, err := e.client.Txn(ctx).Then(deleteOPs...).Commit()
 	if err != nil {
 		return false, err
@@ -89,25 +76,27 @@ func (e *etcdSchemaRegistry) CreateGroup(ctx context.Context, group *commonv1.Gr
 	if group.UpdatedAt != nil {
 		group.UpdatedAt = timestamppb.Now()
 	}
-	return e.create(ctx, Metadata{
+	_, err := e.create(ctx, Metadata{
 		TypeMeta: TypeMeta{
 			Kind: KindGroup,
 			Name: group.GetMetadata().GetName(),
 		},
 		Spec: group,
 	})
+	return err
 }
 
 func (e *etcdSchemaRegistry) UpdateGroup(ctx context.Context, group *commonv1.Group) error {
-	return e.update(ctx, Metadata{
+	_, err := e.update(ctx, Metadata{
 		TypeMeta: TypeMeta{
 			Kind: KindGroup,
 			Name: group.GetMetadata().GetName(),
 		},
 		Spec: group,
 	})
+	return err
 }
 
 func formatGroupKey(group string) string {
-	return groupsKeyPrefix + group + groupMetadataKey
+	return path.Join(groupsKeyPrefix, group)
 }